% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agdist-package.R
\docType{package}
\name{agdist-package}
\alias{agdist}
\alias{agdist-package}
\title{agdist: A package to compute different antigenic distance measures for influenza}
\description{
This package provides a number of methods to compute antigenic distances
between two influenza strains.
}
\details{
To learn more about how to use the package, see the documentation
on the package website:
https://ahgroup.github.io/agdist/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ahgroup.github.io/agdist/index.html}
  \item Report bugs at \url{https://github.com/ahgroup/agdist/issues}
}

}
\author{
\strong{Maintainer}: Andreas Handel \email{andreas.handel@gmail.com} (\href{https://orcid.org/0000-0002-4622-1146}{ORCID})

Authors:
\itemize{
  \item Zane Billings \email{wzbillings@proton.me} (\href{https://orcid.org/0000-0002-0184-6134}{ORCID})
}

}
\keyword{internal}
